Rely on fedmsg-provided queue.
fedbadges implemented this little callLater mechanism a long time ago
before fedmsg and moksha provided a much more sophisticated queueing and
monitoring framework.  This callLater logic actually serves to **hide**
how many messages are in the fedbadges backlog.  By removing it, we can
get a better sense of what the consumer is doing.